requirements.txt
rich>=13.0.0
pyyaml>=6.0
requests>=2.31.0
python-whois>=0.8.0
dnspython>=2.4.0
urllib3>=2.0.0